Consolidate CI-gate scripts into scripts/ci/ under ruff+pyrefly
The scripts invoked directly by .github/workflows were scattered
(scripts/workflows/, scripts/canary/, scripts/datakit/, scripts/*.py) and all
but scripts/workflows/ were excluded from lint and type checking. Gather them
into one home, scripts/ci/, and put the whole directory under ruff + pyrefly so
the gates that decide CI outcomes are themselves linted and type-checked:

  scripts/workflows/iris_monitor.py          -> scripts/ci/iris_monitor.py
  scripts/canary/validate_canary_metrics.py  -> scripts/ci/validate_canary_metrics.py
  scripts/grug_variant_diff_ci.py            -> scripts/ci/grug_variant_diff_ci.py
  scripts/grug_dir_diff.py                   -> scripts/ci/grug_dir_diff.py
  scripts/datakit/collect_perf_metrics.py    -> scripts/ci/collect_perf_metrics.py
  scripts/datakit/submit_perf_run.py         -> scripts/ci/submit_perf_run.py

Pure move + reference update, no behavior change. Updates every caller: the
workflow run/-m/paths references, the inter-script import, the test imports
(tests/workflows/ -> tests/ci/), and the skill/doc paths (triage-canary,
change-grug, evaluate-zephyr-perf, experiments/grug/README). pyproject swaps the
ruff extend-include and pyrefly project-include from scripts/workflows to
scripts/ci and drops the now-empty scripts/canary exclude; scripts/datakit stays
(other tools remain there). ruff + pyrefly are clean on the five newly-covered
files; the four moved-module test files pass (13 tests). Also removes a dead
`scripts/workflows/_open_pr.sh` reference from .github/workflows/README.md.

See #6515.
