+++ .github/workflows/iris-smoke-coreweave.yaml
@@ -6,7 +6,7 @@ on:
     paths:
       - "lib/iris/**"
       - ".github/workflows/iris-smoke-coreweave.yaml"
-      - "scripts/workflows/iris_monitor.py"
+      - "scripts/ci/iris_monitor.py"
   workflow_dispatch:
 
 permissions:
@@ -97,7 +97,7 @@ jobs:
 
       - name: Connect to iris-ci controller
         run: |
-          uv run python scripts/workflows/iris_monitor.py coreweave-controller \
+          uv run python scripts/ci/iris_monitor.py coreweave-controller \
             --namespace "$IRIS_NAMESPACE" \
             --kubeconfig "$HOME/.kube/coreweave-iris" \
             --log-path "$RUNNER_TEMP/iris-cw-port-forward.log"
@@ -148,7 +148,7 @@ jobs:
           fi
           # No job submitted in this lane, so the iris.job_id selector matches
           # nothing — continue-on-error tolerates the empty kubernetes-pods.json.
-          uv run python scripts/workflows/iris_monitor.py collect \
+          uv run python scripts/ci/iris_monitor.py collect \
             --job-id "ci-smoke" \
             --controller-url "$IRIS_CONTROLLER_URL" \
             --provider coreweave \
