+++ .github/workflows/marin-canary-datakit-tier3.yaml
@@ -62,7 +62,7 @@ jobs:
           chmod 644 ~/.ssh/google_compute_engine.pub
 
       # If you change the iris flags below, also update the matching tier
-      # entry in scripts/datakit/submit_perf_run.py (TIERS["3"]) so the
+      # entry in scripts/ci/submit_perf_run.py (TIERS["3"]) so the
       # evaluate-zephyr-perf skill's treatment job stays comparable to this baseline.
       - name: Submit datakit nemotron ferry
         id: submit
@@ -99,7 +99,7 @@ jobs:
       - name: Wait for datakit nemotron ferry
         shell: bash -l {0}
         run: |
-          uv run python scripts/workflows/iris_monitor.py wait \
+          uv run python scripts/ci/iris_monitor.py wait \
             --job-id "${{ steps.submit.outputs.job_id }}" \
             --iris-config "${{ env.IRIS_CONFIG }}" \
             --poll-interval 30 \
@@ -110,7 +110,7 @@ jobs:
         if: success()
         shell: bash -l {0}
         run: |
-          uv run python scripts/datakit/collect_perf_metrics.py \
+          uv run python scripts/ci/collect_perf_metrics.py \
             --job-id "${{ steps.submit.outputs.job_id }}" \
             --iris-config "${{ env.IRIS_CONFIG }}" \
             --status "$FERRY_STATUS_PATH" \
@@ -134,7 +134,7 @@ jobs:
           DIAG_DIR: ${{ github.workspace }}/datakit-nemotron-diagnostics
         run: |
           mkdir -p "$DIAG_DIR"
-          uv run python scripts/workflows/iris_monitor.py collect \
+          uv run python scripts/ci/iris_monitor.py collect \
             --job-id "$JOB_ID" \
             --iris-config "${{ env.IRIS_CONFIG }}" \
             --provider gcp \
