+++ .github/workflows/ops-grug-variant-diff.yaml
@@ -27,8 +27,8 @@ jobs:
           filters: |
             relevant:
               - 'experiments/grug/**'
-              - 'scripts/grug_dir_diff.py'
-              - 'scripts/grug_variant_diff_ci.py'
+              - 'scripts/ci/grug_dir_diff.py'
+              - 'scripts/ci/grug_variant_diff_ci.py'
               - '.github/workflows/ops-grug-variant-diff.yaml'
 
   grug-variant-diff:
@@ -50,7 +50,7 @@ jobs:
 
       - name: Generate variant diff reports
         run: |
-          python -m scripts.grug_variant_diff_ci \
+          python -m scripts.ci.grug_variant_diff_ci \
             --base-sha "${{ github.event.pull_request.base.sha }}" \
             --head-sha "${{ github.event.pull_request.head.sha }}" \
             --output-dir grug-diff-report \
