+++ scripts/ci/validate_canary_metrics.py
